pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hydrakon"
version = "0.1.0"
description = "CLI tool for testing Hydrakon packages"
authors = [
    {name = "Aditya"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer[all]",
    "rich",
    "PyYAML"
]

[project.scripts]
hydrakon = "hydrakon.main:app"
hdk = "hydrakon.main:app"

[tool.setuptools]
packages = ["hydrakon"]